requirements.txt
google-api-python-client==1.6.5
gcloud==0.18.3
google-cloud-bigquery==1.5.1
